Prepare for F register conversion and opcode grouping
Prepare for migration to using a uint8_t for the flags by adding a
lookup table for the S, Z, P and R flags.  The R flag refers to
bit 1, which is always 1.

Prepare for opcode grouping by creating an enum with the groups of
opcodes can be processed using common code.