Migrate MOV register,registser
Add an array of pointers to 8-bit registers that corresponds to the
bits in the opcode:

000 = B
001 = C
010 = D
011 = E
100 = H
101 = L
110 = N/A
111 = A

Initialise the array in i8080_init().  Create new functions DDD and
SSS that can be used to extract the source and destination register
index to the table from the opcode.

In i8080_execute, lookup the instruction group from the lookup table
and if it is MOV_R_R, use the new code.  Remove the code from the
old switch statement.

We have now migrated the first instruction group.