Migrate LXI RR,NN
Add an array of pointers to 16-bit registers that corresponds to the
bits in the opcode:

00 = BC
01 = DE
10 = HL
11 = SP

Initialise the array in i8080_init().  Create new function RP that
can be used to lookup and index to the table from the opcode.

Migrate the LXI RR,NN opcodes to use the new capability.